LibC+LibRuntime: Mostly mechanically move FILE implementation
The idea is to provide a few functions like fopen, fwrite, and fclose in
LibRuntime to allow seamless interoperability between LibC output and
AK's output functions. Whether we need this interoperability is
debatable but I decided to not change existing behavior for now.

The commit itself would have been trivial cut-paste (since it doesn't
yet port anything to LibRuntime wrappers) but for fflush and
fclose sharing static variables with FILE implementation.
