+++ Userland/Libraries/LibC/bits/stdio_file_implementation.h
@@ -27,6 +27,8 @@ public:
     ~FILE();
 
     static FILE* create(int fd, int mode);
+    static int close(FILE* stream);
+    static int flush_open_streams();
 
     void setbuf(u8* data, int mode, size_t size) { m_buffer.setbuf(data, mode, size); }
